+++ app/src/main/java/de/nproth/pin/pinboard/PinboardService.java
@@ -13,9 +13,11 @@ import de.nproth.pin.receiver.SnoozeNoteReceiver;
 public class PinboardService extends Service {
 
     public static final String INTENT_ACTION_SNOOZE_PIN = BuildConfig.APPLICATION_ID + ".intent.action.snooze";
+    public static final String INTENT_ACTION_WAKE_UP = BuildConfig.APPLICATION_ID + ".intent.action.wake_up";
 
     public static final String PREFERENCE_PERSISTENT_NOTIFICATIONS = "pref_persistent_notifications";
     public static final String PREFERENCE_SNOOZE_DURATION = "snooze_duration";
+
     public static final long DEFAULT_SNOOZE_DURATION = 30 * 60 * 1000; //30min in millis
     public static final boolean DEFAULT_PERSISTENT_NOTIFICATIONS = false;
 
@@ -24,8 +26,12 @@ public class PinboardService extends Service {
 
     public class PinboardBinder extends Binder {
 
+        public void update(boolean silent) {
+            mPinboard.updateAll(silent);
+        }
+
         public void update() {
-            mPinboard.updateVisible();
+            update(true);
         }
 
         public long getSnoozeDuration() {
@@ -57,7 +63,7 @@ public class PinboardService extends Service {
             int rows = SnoozeNoteReceiver.onSnoozePins(this, mPinboard.getSnoozeDuration(), intent.getData());
         }
 
-        mPinboard.updateChanged();
+        mPinboard.updateAll(! INTENT_ACTION_WAKE_UP.equals(intent.getAction()));
         return START_STICKY;
     }
 
@@ -78,6 +84,7 @@ public class PinboardService extends Service {
     public void onDestroy() {
         super.onDestroy();
 
+        //Committing the changed preference values is deliberately done synchronously (blocking), so that the Service is not destroyed before the transaction is completed.
         mPrefs.edit().putLong(PREFERENCE_SNOOZE_DURATION, mPinboard.getSnoozeDuration())
                 .putBoolean(PREFERENCE_PERSISTENT_NOTIFICATIONS, mPinboard.getIsFixed()).commit();
     }